refactor raster layers
