+++ folium/plugins/__init__.py
@@ -17,7 +17,6 @@ from folium.plugins.float_image import FloatImage
 from folium.plugins.fullscreen import Fullscreen
 from folium.plugins.heat_map import HeatMap
 from folium.plugins.heat_map_withtime import HeatMapWithTime
-from folium.plugins.image_overlay import ImageOverlay
 from folium.plugins.marker_cluster import MarkerCluster
 from folium.plugins.measure_control import MeasureControl
 from folium.plugins.polyline_text_path import PolyLineTextPath
@@ -25,7 +24,6 @@ from folium.plugins.scroll_zoom_toggler import ScrollZoomToggler
 from folium.plugins.terminator import Terminator
 from folium.plugins.timestamped_geo_json import TimestampedGeoJson
 from folium.plugins.timestamped_wmstilelayer import TimestampedWmsTileLayers
-from folium.plugins.video_overlay import VideoOverlay
 
 __all__ = [
     'BoatMarker',
@@ -35,7 +33,6 @@ __all__ = [
     'Fullscreen',
     'HeatMap',
     'HeatMapWithTime',
-    'ImageOverlay',
     'MarkerCluster',
     'MeasureControl',
     'PolyLineTextPath',
@@ -43,5 +40,4 @@ __all__ = [
     'Terminator',
     'TimestampedGeoJson',
     'TimestampedWmsTileLayers',
-    'VideoOverlay',
     ]
