+++ folium/plugins/heat_map_withtime.py
@@ -3,7 +3,7 @@
 from branca.element import CssLink, Element, Figure, JavascriptLink
 from branca.utilities import none_max, none_min
 
-from folium.map import TileLayer
+from folium.raster_layers import TileLayer
 
 from jinja2 import Template
