+++ folium/plugins/timestamped_wmstilelayer.py
@@ -4,7 +4,7 @@ from __future__ import (absolute_import, division, print_function)
 
 from branca.element import CssLink, Figure, JavascriptLink
 
-from folium.features import WmsTileLayer
+from folium.raster_layers import WmsTileLayer
 from folium.map import Layer
 
 from jinja2 import Template
